% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_constraints.R
\name{constraint_function}
\alias{constraint_function}
\title{Constraint, Direction and RHS functions all for defining the boundaries and constraints for the optimisation}
\usage{
constraint_function(input)
}
\description{
Used largely behind the scenes within other denoted optimisation functions
}
