% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_calculate_predictors.R
\name{fpl_calculate_predictors}
\alias{fpl_calculate_predictors}
\title{Calculate each of the unique indexes for modelling}
\usage{
fpl_calculate_predictors(
  players = players,
  period,
  weight = 0.5,
  strength_index = 1,
  odds_gs = odds_gs_gw8,
  odds_cs = odds_cs_gw8
)
}
\arguments{
\item{period}{Fixtures taking place in the specified period.}

\item{weight}{Adjustment to weight predictions based more heavily on either current form or historical performance.}

\item{strength_index}{"Low", "Med" or "High" to define influence fixtures have on our expected points.}

\item{odds_gs}{Goalscoring and Assists odds for current week.}

\item{odds_cs}{Odds for each team to keep a clean sheet this week.}
}
\value{
Dataframe
}
\description{
Converts fields such as average points, ict rating etc. into usable indexes for generating predicted points
}
