% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_calculate_xP.R
\name{fpl_calculate_xP}
\alias{fpl_calculate_xP}
\title{Calculate expected points for each player}
\usage{
fpl_calculate_xP(
  data = players_index,
  predictors = predictors_indexes,
  user,
  gw = 8
)
}
\arguments{
\item{data}{Output from "fpl_calculate_predictors" function used as the initial starting point}

\item{predictors}{select which variables to use as predicting variables - either odds or indexes}
}
\value{
Dataframe.
}
\description{
Two separate methods - use odds data, or use other indexes (average points, ict index etc.)
}
