% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_fixtures_difficulty_rating.R
\name{fpl_fixtures_difficulty_rating}
\alias{fpl_fixtures_difficulty_rating}
\title{Produce a table detailing fixture difficulty for each team over future gameweeks.}
\usage{
fpl_fixtures_difficulty_rating(gameweek1, gameweek2)
}
\arguments{
\item{gameweek1}{Period start date.}

\item{gameweek2}{Period end date.}
}
\description{
Investigate which teams have the easiest and hardest run of fixtures coming up.
}
\examples{

}
