% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_gameweek_predictions.R
\name{fpl_gameweek_predictions}
\alias{fpl_gameweek_predictions}
\title{Return expected points predictions for a specified gameweek}
\usage{
fpl_gameweek_predictions(players = player_details, gameweek = 8)
}
\arguments{
\item{gameweek}{Gameweek for which to return predictions for}

\item{player_details}{The result of fplscrapR::get_player_details()}

\item{predictions}{"odds" or "index" - Decide which method to call}
}
\value{
Dataframe.
}
\description{
Master function compiling outputs from various other functions.
}
