% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_get_api_response.R
\name{fpl_get_api_response}
\alias{fpl_get_api_response}
\title{Retrieve API data directly from the fpl website}
\usage{
fpl_get_api_response(key, secret, user)
}
\arguments{
\item{key}{Email for your fpl account}

\item{user}{Unique user id}

\item{secert}{Password for fpl account}
}
\description{
Requires username and password authentication to retrieve more granular information
}
