% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_load_historic_data.R
\name{fpl_load_historic_data}
\alias{fpl_load_historic_data}
\title{Load historic gameweek data from the 2021/22 season}
\usage{
fpl_load_historic_data()
}
\value{
Dataframe.
}
\description{
This data offers a gameweek by gameweek breakdown of fields such as total_points, ICT score, etc.
}
