% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_my_budget.R
\name{fpl_my_budget}
\alias{fpl_my_budget}
\title{Find the budegt available for transfers}
\usage{
fpl_my_budget(user, key, secret)
}
\arguments{
\item{user}{Unique identifier}

\item{key}{fpl account username}

\item{secret}{fpl account password}
}
\value{
Integer.
}
\description{
Calculaed via the selling price column returned from the "fpl_get_api_response()" function
}
