% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_my_team.R
\name{fpl_my_team}
\alias{fpl_my_team}
\title{Return a list containing all players selected in your team}
\usage{
fpl_my_team(user, squad = 11, gw = 4)
}
\arguments{
\item{user}{Unique identifier}

\item{squad}{Either 11 or 15, depending on whether to select all team including or excluding subs}

\item{gw}{Numeric value, should be set to the most recently passed gameweek (gw4 at time of writing)S}
}
\description{
Converts fields such as average points, ict rating etc. into usable indexes for generating predicted points
}
