% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_optimise.R
\name{fpl_optimise}
\alias{fpl_optimise}
\title{Select the top 11 players subject to expected points, and boundable constraints}
\usage{
fpl_optimise(input = players_xP, obj_var = players_xP$xP, budget = 83)
}
\arguments{
\item{input}{Output from "fpl_calculate_xP()" function}

\item{obj_var}{Variable to optimise based on (xP by default)}

\item{budget}{total funds available to spend on first 11 players}
}
\value{
Solution to optimisation problem.
}
\description{
Constraints are applicable for positions, team, budget, etc.
}
