% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_optimise_my_team.R
\name{fpl_optimise_my_team}
\alias{fpl_optimise_my_team}
\title{Decide which transfers should be made for the users' team with this function}
\usage{
fpl_optimise_my_team(input, obj_var, budget = 83, transfers = 2)
}
\arguments{
\item{input}{Output from "fpl_calculate_xP()" function}

\item{obj_var}{Variable to optimise based on (xP by default)}

\item{budget}{total funds available to spend on first 11 players}

\item{transfers}{Number of free transfers available to the user}
}
\description{
Transfers argument is flexible to denote how many changes can be made this week
}
