% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_performance_comparison.R
\name{fpl_performance_comparison}
\alias{fpl_performance_comparison}
\title{Compare how predictions shape up against observed actuals, one gameweek at a time.}
\usage{
fpl_performance_comparison(players, gw = "gw4")
}
\arguments{
\item{gw}{Gameweek we are comparing}
}
\value{
Dataframe.
}
\description{
Print correlation for both methods for a side by side comparison
}
