% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{rsq}
\alias{rsq}
\title{rsq, range01 and str_convert - functions used mid process inside other functions within the fplBuddy package.}
\usage{
rsq(x, y)
}
\description{
rsq, range01 and str_convert - functions used mid process inside other functions within the fplBuddy package.
}
