% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_fixtures.R
\name{fpl_fixtures}
\alias{fpl_fixtures}
\title{Return a dataframe containing all fixtures coming up between a start and end date.}
\usage{
fpl_fixtures(gameweek1, gameweek2)
}
\arguments{
\item{gameweek1}{Period start date.}

\item{gameweek2}{Period end date.}
}
\value{
Dataframe
}
\description{
Denotes home and away team strength for each team in each fixture.
}
