% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_odds_generator_cs.R
\name{fpl_odds_generator_cs}
\alias{fpl_odds_generator_cs}
\title{Return the odds for each team to keep a cleansheet for the next gameweek.}
\usage{
fpl_odds_generator_cs(url, xpath)
}
\arguments{
\item{url}{Web link where odds table is scraped from}

\item{xpath}{Unique identifier for the clean sheet probability table}
}
\description{
Return the odds for each team to keep a cleansheet for the next gameweek.
}
