% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpl_odds_generator_gs.R
\name{fpl_odds_generator_gs}
\alias{fpl_odds_generator_gs}
\title{Return the odds for each player to score or register an assist for the next gameweek.}
\usage{
fpl_odds_generator_gs(data = odds_gs_gw5)
}
\arguments{
\item{data}{Latest odds data for likelihood of scoring and assisting - gw5 at the time of writing}
}
\description{
Return the odds for each player to score or register an assist for the next gameweek.
}
